refactor(bucket): separate files
