+++ src/search/bucket/relay.go
@@ -1,0 +1,12 @@
+package bucket
+
+import (
+	"sync"
+
+	"github.com/hearchco/hearchco/src/search/result"
+)
+
+type Relay struct {
+	ResultMap map[string]*result.Result
+	Mutex     sync.RWMutex
+}
